server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.13.0",
    "argon2-cffi>=23.1.0",
    "asyncpg>=0.29.0",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "openai-agents>=0.6.4",
    "pydantic[email]>=2.12.5",
    "pyjwt[crypto]>=2.8.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.9",
    "qdrant-client>=1.16.2",
    "requests>=2.32.5",
    "slowapi>=0.1.9",
    "sqlalchemy[asyncio]>=2.0.0",
    "uvicorn>=0.40.0",
]
